pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lufthansa-inspiria-strands"
version = "0.1.0"
description = "Lufthansa Inspiria agents powered by AWS Strands"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
dependencies = [
    "strands-agents>=1.16.0",
    "strands-agents-tools>=0.2.15",
    "pydantic>=2.9",
    "pydantic-settings>=2.6.1",
    "opentelemetry-exporter-otlp>=1.38.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.8",
    "ruff>=0.6",
    "pytest>=7.4",
]

[tool.black]
line-length = 100
target-version = ["py310"]
extend-exclude = '''
/(
    vendor
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py310"
line-length = 100
extend-exclude = [
    "vendor",
    "**/build",
    "**/dist",
    "**/.venv",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "UP", # pyupgrade
    "C90",# mccabe
]

[tool.pytest.ini_options]
addopts = "-q -s"
testpaths = ["tests"]
